Matrix dimensions off

+++ supervised_learning/0x01-classification/16-deep_neural_network.py
@@ -32,6 +32,5 @@ class DeepNeuralNetwork:
             l1 = layers[i - 1]
             w = 'W' + str(i)
             b = 'b' + str(i)
-            self.weights.update({w: np.random.randn(l1,
-                                int(l2 * np.sqrt(2 / l2)))})
+            self.weights.update({w: np.random.randn(l1, l2) * np.sqrt(2 / l2)})
             self.weights.update({b: np.zeros((l1, 1))})
